fix(StoryItem): uses the right param for story item create/update api call
When the controller name is not used to wrap the params that are send on a post/put method
ParamsWrapper makes a copy of the params and adds them to params with a new key named after the controller name.

This clutters the logs as the params are doubled. Fix is to follow Rails convention and used the same
name as the controller for params.